+++ noodles-util/examples/util_variant_view.rs
@@ -31,7 +31,7 @@ fn main() -> io::Result<()> {
 
     for result in reader.records(&header) {
         let record = result?;
-        writer.write_record(&record)?;
+        writer.write_record(&header, &record)?;
     }
 
     Ok(())